requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-WTF==0.15.1
requests==2.26.0
python-dotenv==0.19.0
WTForms==2.3.3
flask-cors==3.0.10
eventlet==0.33.0
